Refactor PoC code into namespaces, add led management, improve directory structure
