+++ ESP/include/WifiHandler.h
@@ -1,0 +1,6 @@
+#include <WiFi.h>
+#include "pinout.h"
+
+namespace WiFiHandler {
+    void setupWifi();
+}
