+++ ESP/platformio.ini
@@ -17,3 +17,6 @@ monitor_rts = 0
 monitor_dtr = 0
 build_flags = -O2
 build_unflags = -Os
+
+lib_deps =
+  esp32-camera
