--- OpenEyeTrackingESP/src/httpdHandler.cpp
@@ -1,185 +1,0 @@
-#include <Arduino.h>
-#include "esp_http_server.h"
-#include "esp_camera.h"
-
-#define PART_BOUNDARY "123456789000000000000987654321"
-
-httpd_handle_t camera_httpd = NULL;
-httpd_handle_t control_httpd = NULL;
-
-static const char *_STREAM_CONTENT_TYPE = "multipart/x-mixed-replace;boundary=" PART_BOUNDARY;
-static const char *_STREAM_BOUNDARY = "\r\n--" PART_BOUNDARY "\r\n";
-static const char *_STREAM_PART = "Content-Type: image/jpeg\r\nContent-Length: %u\r\nX-Timestamp: %d.%06d\r\n\r\n";
-
-
-static esp_err_t stream_handler(httpd_req_t *req){
-    camera_fb_t *fb = NULL;
-    struct timeval _timestamp;
-
-    esp_err_t res = ESP_OK;
-
-    size_t _jpg_buf_len = 0;
-    uint8_t *_jpg_buf = NULL;
-
-    char *part_buf[128];
-    
-    static int64_t last_frame = 0;
-    if (!last_frame)
-        last_frame = esp_timer_get_time();
-
-    res = httpd_resp_set_type(req, _STREAM_CONTENT_TYPE);
-    if (res != ESP_OK)
-        return res;
-
-    httpd_resp_set_hdr(req, "Access-Control-Allow-Origin; Content-Type: multipart/x-mixed-replace; boundary=123456789000000000000987654321\r\n", "*");
-    httpd_resp_set_hdr(req, "X-Framerate", "60");
-
-    while (true) {
-        fb = esp_camera_fb_get();
-        if (!fb){
-            ESP_LOGE(TAG, "Camera capture failed");
-            res = ESP_FAIL;
-        }
-        else{
-            _timestamp.tv_sec = fb->timestamp.tv_sec;
-            _timestamp.tv_usec = fb->timestamp.tv_usec;
-                if (fb->format != PIXFORMAT_JPEG){
-                    bool jpeg_converted = frame2jpg(fb, 80, &_jpg_buf, &_jpg_buf_len);
-                    esp_camera_fb_return(fb);
-                    fb = NULL;
-                    if (!jpeg_converted){
-                        ESP_LOGE(TAG, "JPEG compression failed");
-                        res = ESP_FAIL;
-                    }
-                }
-                else{
-                    _jpg_buf_len = fb->len;
-                    _jpg_buf = fb->buf;
-                }
-        }
-        if (res == ESP_OK){
-            res = httpd_resp_send_chunk(req, _STREAM_BOUNDARY, strlen(_STREAM_BOUNDARY));
-        }
-        if (res == ESP_OK){
-            size_t hlen = snprintf((char *)part_buf, 128, _STREAM_PART, _jpg_buf_len, _timestamp.tv_sec, _timestamp.tv_usec);
-            res = httpd_resp_send_chunk(req, (const char *)part_buf, hlen);
-        }
-        if (res == ESP_OK){
-            res = httpd_resp_send_chunk(req, (const char *)_jpg_buf, _jpg_buf_len);
-        }
-        if (fb){
-            esp_camera_fb_return(fb);
-            fb = NULL;
-            _jpg_buf = NULL;
-        }
-        else if (_jpg_buf){
-            free(_jpg_buf);
-            _jpg_buf = NULL;
-        }
-        if (res != ESP_OK){
-            break;
-        }
-    }
-    last_frame = 0;
-    return res;
-}
-
-static esp_err_t parse_get(httpd_req_t *req, char **obuf)
-{
-    char *buf = NULL;
-    size_t buf_len = 0;
-
-    buf_len = httpd_req_get_url_query_len(req) + 1;
-    if (buf_len > 1) {
-        buf = (char *)malloc(buf_len);
-        if (!buf) {
-            httpd_resp_send_500(req);
-            return ESP_FAIL;
-        }
-        if (httpd_req_get_url_query_str(req, buf, buf_len) == ESP_OK) {
-            *obuf = buf;
-            return ESP_OK;
-        }
-        free(buf);
-    }
-    httpd_resp_send_404(req);
-    return ESP_FAIL;
-}
-
-static esp_err_t command_handler(httpd_req_t *req) {
-    char *buf = NULL;
-    char variable[32];
-    char value[32];
-
-    if (parse_get(req, &buf) != ESP_OK) {
-        return ESP_FAIL;
-    }
-    if (httpd_query_key_value(buf, "var", variable, sizeof(variable)) != ESP_OK ||
-        httpd_query_key_value(buf, "val", value, sizeof(value)) != ESP_OK) {
-        free(buf);
-        httpd_resp_send_404(req);
-        return ESP_FAIL;
-    }
-    free(buf);
-
-    // debug
-    Serial.printf("%s %s \r\n", variable, value);
-
-    sensor_t *s = esp_camera_sensor_get();
-    int val = atoi(value);
-    ESP_LOGI(TAG, "%s = %d", variable, val);
-    int res = 0;
-
-    if (!strcmp(variable, "framesize")) {
-        if (s->pixformat == PIXFORMAT_JPEG) {
-            res = s->set_framesize(s, (framesize_t)val);
-        }
-    }
-    else if (!strcmp(variable, "hmirror"))
-        res = s->set_hmirror(s, val);
-    else if (!strcmp(variable, "vflip"))
-        res = s->set_vflip(s, val);
-    else {
-        // invalid command
-        res = -1;
-    }
-
-    if (res < 0){
-        return httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, "Command not supported");
-    }
-
-    httpd_resp_set_hdr(req, "Access-Control-Allow-Origin", "*");
-    return httpd_resp_send(req, NULL, 0);
-}
-
-void startStreamServer(){
-    Serial.println("Setting up the server");
-
-    httpd_config_t config = HTTPD_DEFAULT_CONFIG();
-    config.max_uri_handlers = 3;
-
-    httpd_uri_t control_page = {
-        .uri = "/control/",
-        .method = HTTP_GET,
-        .handler = &command_handler,
-        .user_ctx = NULL
-    };
-
-    httpd_uri_t stream_page = {
-        .uri = "/",
-        .method = HTTP_GET,
-        .handler = &stream_handler,
-        .user_ctx = NULL
-    };
-
-    if (httpd_start(&control_httpd, &config) == ESP_OK)
-        httpd_register_uri_handler(control_httpd, &control_page);
-
-    config.server_port += 1;
-    config.ctrl_port += 1;
-
-    if (httpd_start(&camera_httpd, &config) == ESP_OK)
-        httpd_register_uri_handler(camera_httpd, &stream_page);
-
-    Serial.println("Server is ready to serve!");
-}
