+++ docs/.obsidian/core-plugins.json
