+++ docs/.obsidian/workspace
