+++ docs/Goal and project description.md
