+++ docs/Hardware.md
